requirements.txt
openai>=1.30.0
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
tqdm>=4.66.0
scikit-learn>=1.4.0
pypdf>=4.0.0
python-docx>=1.1.0
